app/searcher/requirements.txt
fastapi
uvicorn
pydantic
httpx
pymongo
python-dotenv
python-multipart